Updated userNFT to collection, extracted components logic
