+++ frontend/src/components/Actions/Actions.tsx
@@ -11,7 +11,7 @@ const Actions: React.FC<ActionProps> = ({ userAddress }) => {
     <>
       <div className="col-8 offset-2 pt-5 pb-5 mt-5">
         <h1 className="fw-light">Start Playing</h1>
-        <Link to={`/NFT/user/${userAddress}`}>
+        <Link to={`/Collection/${userAddress}`}>
           <button className="btn btn-primary btn-lg mb-4">Play</button>
         </Link>
       </div>
