+++ frontend/src/components/Misc/LoadingBtn.tsx
@@ -1,6 +1,6 @@
 // *EXTERNALS*
 import React from 'react';
-import Button from '../Button';
+import Button from './Button';
 
 interface BtnProps {
   type?: string;
