+++ frontend/src/routes/SingleLOOT.tsx
@@ -2,6 +2,8 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
 import { Buffer } from 'buffer';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faCircleLeft } from '@fortawesome/free-solid-svg-icons';
 
 // *INTERNALS*
 import { useTokenURILoot } from '../hooks/index';
@@ -15,8 +17,11 @@ export default function SingleLOOT() {
   if (LOOTObject) LOOTObject.tokenId = lootId;
 
   return (
-    <div className="text-center mt-5">
+    <div className="mt-5">
       <div className="container container-decorated col-4" style={{ minHeight: '600px' }}>
+        <a href="#" onClick={() => history.back()}>
+          <FontAwesomeIcon icon={faCircleLeft} fontSize={25} />
+        </a>
         <div className="col-8 offset-2">{LOOTObject && <CardWrapper NFT={LOOTObject} isLoot />}</div>
       </div>
     </div>
