+++ frontend/src/routes/UserLOOT.tsx
@@ -14,11 +14,11 @@ const fetcher = (params: any) => fetch(params).then((res) => res.json());
 export default function UserLoot() {
   const params = useParams();
   const { userAddress } = params;
-  const { data: userLOOT, error } = useSWR(`${API_ADDRESS}/v1/loot/wallet/${userAddress}`, fetcher);
+  const { data: userLOOT } = useSWR(`${API_ADDRESS}/v1/loot/wallet/${userAddress}`, fetcher);
 
   return (
     <>
-      <h2 className="mt-5 mb-2">Collection</h2>
+      <h2 className="mt-5 mb-2">Loot</h2>
       {userLOOT && userLOOT.length > 0 && userAddress ? (
         <div className="row row-cols-1 row-cols-sm-2 row-cols-md-3 g-3">
           {userLOOT.map((NFT: NFT) => (
